% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE1910_metazoan_Dec2022.R
\name{AE1910_metazoan_Dec2022}
\alias{AE1910_metazoan_Dec2022}
\title{Proteomics NMDS and hierarchical clustering}
\usage{
AE1910_metazoan_Dec2022()
}
\value{

}
\description{
Proteomics NMDS and hierarchical clustering
}
