% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_transcriptomics_Dec2022.R
\name{DAPC_transcriptomics_Dec2022}
\alias{DAPC_transcriptomics_Dec2022}
\title{Discriminant analysis of principal components (DAPC) of transcriptomics dataset}
\usage{
DAPC_transcriptomics_Dec2022()
}
\value{

}
\description{
Discriminant analysis of principal components (DAPC) of transcriptomics dataset
}
