% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_NMDS_Dec2022.R
\name{transcriptomics_NMDS_Dec2022}
\alias{transcriptomics_NMDS_Dec2022}
\title{Nonmetric multidimensional scaling (NMDS) of transcriptomics dataset}
\usage{
transcriptomics_NMDS_Dec2022()
}
\value{

}
\description{
Nonmetric multidimensional scaling (NMDS) of transcriptomics dataset
}
